requirements.txt
asgiref==3.7.2
Django==4.2.1
django-cleanup==7.0.0
django-colorfield==0.8.0
django-filter==23.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
Markdown==3.4.3
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
